requirements.txt
# generated from manifests external_dependencies
cryptography
extendable-pydantic
pydantic==1.10.10
python-jose
